server/requirements.txt
alembic==1.7.7
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
mariadb==1.0.10
MarkupSafe==2.1.0
mysqlclient==2.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
